Update unit-2.md
One of the other pages about learning activities for this unit gave a lot of detail about AI and the activity, but none about activity 2.2 regarding the pressbooks.  Not sure if mention should be made there as well about 2.2. I actually removed reference to 2.2 because there was no further mention of it on that page.